Finding area using a Monte-Carlo approach

Finding area using a Monte-Carlo approach

estimate_area(B = 5000, seed = 10)

Arguments

B
number of points

seed
for random number

Value

Return the estimated area and the coordinates of the points

Author

Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger

Examples

estimate_area(50)
#> $estimated_area
#> [1] 0.6
#> 
#> $points
#>             x          y inside
#> 1  0.50747820 0.35432806   TRUE
#> 2  0.30676851 0.93643254   TRUE
#> 3  0.42690767 0.24586639  FALSE
#> 4  0.69310208 0.47314146   TRUE
#> 5  0.08513597 0.19156087  FALSE
#> 6  0.22543662 0.58322197   TRUE
#> 7  0.27453052 0.45947319   TRUE
#> 8  0.27230507 0.46743405   TRUE
#> 9  0.61582931 0.39983256   TRUE
#> 10 0.42967153 0.50528560   TRUE
#> 11 0.65165567 0.03188816  FALSE
#> 12 0.56773775 0.11446759   TRUE
#> 13 0.11350898 0.46893548  FALSE
#> 14 0.59592531 0.39698674   TRUE
#> 15 0.35804998 0.83361919   TRUE
#> 16 0.42880942 0.76112174   TRUE
#> 17 0.05190332 0.57335645   TRUE
#> 18 0.26417767 0.44750805   TRUE
#> 19 0.39879073 0.08380201  FALSE
#> 20 0.83613414 0.21913855  FALSE
#> 21 0.86472123 0.07557029  FALSE
#> 22 0.61535242 0.53442678   TRUE
#> 23 0.77510990 0.64135658   TRUE
#> 24 0.35556869 0.52573932   TRUE
#> 25 0.40584997 0.03928139  FALSE
#> 26 0.70664691 0.54585984   TRUE
#> 27 0.83828767 0.37276310   TRUE
#> 28 0.23958913 0.96130241  FALSE
#> 29 0.77077153 0.25734157  FALSE
#> 30 0.35589774 0.20795168  FALSE
#> 31 0.53559704 0.86138244   TRUE
#> 32 0.09308813 0.46439198  FALSE
#> 33 0.16980304 0.22286743  FALSE
#> 34 0.89983245 0.62354960   TRUE
#> 35 0.42263761 0.20364770  FALSE
#> 36 0.74774647 0.01967341  FALSE
#> 37 0.82265258 0.79799301   TRUE
#> 38 0.95465365 0.27431890  FALSE
#> 39 0.68544451 0.16660910  FALSE
#> 40 0.50050323 0.17015172   TRUE
#> 41 0.27548386 0.48850594   TRUE
#> 42 0.22890394 0.71140908   TRUE
#> 43 0.01443391 0.59193492   TRUE
#> 44 0.72896456 0.51715688   TRUE
#> 45 0.24988047 0.38162769  FALSE
#> 46 0.16118328 0.83477833   TRUE
#> 47 0.01704265 0.24900805  FALSE
#> 48 0.48610035 0.54060103   TRUE
#> 49 0.10290017 0.74383940   TRUE
#> 50 0.80154700 0.03001186  FALSE
#> 
#> attr(,"class")
#> [1] "area"